+++ src/main/java/com/example/HotelManagment/Model/Employee.java
@@ -12,7 +12,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long employeeId;
+    private int  employeeId;
 
     private String firstName;
     private String lastName;
@@ -34,11 +34,11 @@ public class Employee {
     }
 
     // Getters and Setters
-    public Long getEmployeeId() {
+    public int getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(Long employeeId) {
+    public void setEmployeeId(int employeeId) {
         this.employeeId = employeeId;
     }
