+++ src/main/java/com/example/HotelManagment/Repo/BookingRepository.java
@@ -3,5 +3,5 @@ package com.example.HotelManagment.Repo;
 import com.example.HotelManagment.Model.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<Booking, Long> {}
+public interface BookingRepository extends JpaRepository<Booking, Integer> {}
