+++ src/main/java/com/example/HotelManagment/Repo/EmployeeRepository.java
@@ -3,5 +3,5 @@ package com.example.HotelManagment.Repo;
 import com.example.HotelManagment.Model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {}
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {}
