+++ src/main/java/com/example/HotelManagment/Repo/GuestRepository.java
@@ -3,4 +3,4 @@ package com.example.HotelManagment.Repo;
 import com.example.HotelManagment.Model.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuestRepository extends JpaRepository<Guest, Long> {}
+public interface GuestRepository extends JpaRepository<Guest, Integer> {}
