+++ src/main/java/com/example/HotelManagment/Repo/RoomRepository.java
@@ -3,5 +3,5 @@ package com.example.HotelManagment.Repo;
 import com.example.HotelManagment.Model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Long> {}
+public interface RoomRepository extends JpaRepository<Room, Integer> {}
